pipeline/requirements.txt
python-dotenv>=1.0.1
pandas>=2.0
numpy>=1.24
tqdm>=4.66

scikit-learn>=1.3
hdbscan>=0.8.33
transformers>=4.44
tokenizers>=0.19
torch>=2.3 ; platform_system != "Darwin" or platform_machine != "arm64"

sentencepiece>=0.1.99
protobuf>=3.20,<5
tiktoken>=0.7

voyageai>=0.2.4
anthropic>=0.31.2
jsonschema>=4.22

langdetect>=1.0.9
huggingface_hub[hf_xet]>=0.24

pyarrow>=15.0  # per cache parquet
